refactor: extract server-settings.js from server.js (PR-18)
Move user profile and settings routes into dedicated module:
GET/PUT profile, avatar upload/serve (user and mate), PIN change,
auto-continue toggle, chat layout, mate onboarding flag.

server.js delegates settings routes via settings.handleRequest().

server.js: 2199 -> 1752 lines. server-settings.js: 479 lines.
