+++ lib/server-settings.js
@@ -1,0 +1,479 @@
+var fs = require("fs");
+var path = require("path");
+
+function attachSettings(ctx) {
+  var users = ctx.users;
+  var mates = ctx.mates;
+  var getMultiUserFromReq = ctx.getMultiUserFromReq;
+  var projects = ctx.projects;
+  var opts = ctx.opts;
+  var CONFIG_DIR = ctx.CONFIG_DIR;
+
+  var profilePath = path.join(CONFIG_DIR, "profile.json");
+
+  function handleRequest(req, res, fullUrl) {
+    // GET /api/profile
+    if (req.method === "GET" && fullUrl === "/api/profile") {
+      if (users.isMultiUser()) {
+        var mu = getMultiUserFromReq(req);
+        if (!mu) {
+          res.writeHead(401, { "Content-Type": "application/json" });
+          res.end('{"error":"unauthorized"}');
+          return true;
+        }
+        var profile = mu.profile || { name: "", lang: "en-US", avatarColor: "#7c3aed", avatarStyle: "thumbs", avatarSeed: "", avatarCustom: "" };
+        profile.username = mu.username;
+        profile.userId = mu.id;
+        profile.role = mu.role;
+        profile.autoContinueOnRateLimit = !!mu.autoContinueOnRateLimit;
+        profile.chatLayout = mu.chatLayout || "channel";
+        profile.mateOnboardingShown = !!mu.mateOnboardingShown;
+        res.writeHead(200, { "Content-Type": "application/json" });
+        res.end(JSON.stringify(profile));
+        return true;
+      }
+      var profile = { name: "", lang: "en-US", avatarColor: "#7c3aed", avatarStyle: "thumbs", avatarSeed: "", avatarCustom: "" };
+      try {
+        var raw = fs.readFileSync(profilePath, "utf8");
+        var saved = JSON.parse(raw);
+        if (saved.name !== undefined) profile.name = saved.name;
+        if (saved.lang) profile.lang = saved.lang;
+        if (saved.avatarColor) profile.avatarColor = saved.avatarColor;
+        if (saved.avatarStyle) profile.avatarStyle = saved.avatarStyle;
+        if (saved.avatarSeed) profile.avatarSeed = saved.avatarSeed;
+        if (saved.avatarCustom) profile.avatarCustom = saved.avatarCustom;
+      } catch (e) { /* file doesn't exist yet */ }
+      // Single-user settings from daemon config
+      if (typeof opts.onGetDaemonConfig === "function") {
+        var dc = opts.onGetDaemonConfig();
+        profile.autoContinueOnRateLimit = !!dc.autoContinueOnRateLimit;
+        profile.chatLayout = dc.chatLayout || "channel";
+        profile.mateOnboardingShown = !!dc.mateOnboardingShown;
+      }
+      // Check if custom avatar file exists
+      try {
+        var avatarFiles = fs.readdirSync(path.join(CONFIG_DIR, "avatars"));
+        for (var afi = 0; afi < avatarFiles.length; afi++) {
+          if (avatarFiles[afi].startsWith("default.")) {
+            profile.avatarCustom = "/api/avatar/default?v=" + fs.statSync(path.join(CONFIG_DIR, "avatars", avatarFiles[afi])).mtimeMs;
+            break;
+          }
+        }
+      } catch (e) {}
+      res.writeHead(200, { "Content-Type": "application/json" });
+      res.end(JSON.stringify(profile));
+      return true;
+    }
+
+    // PUT /api/profile
+    if (req.method === "PUT" && fullUrl === "/api/profile") {
+      var body = "";
+      req.on("data", function (chunk) { body += chunk; });
+      req.on("end", function () {
+        try {
+          var data = JSON.parse(body);
+          var profile = {};
+          if (typeof data.name === "string") profile.name = data.name.substring(0, 50);
+          if (typeof data.lang === "string") profile.lang = data.lang.substring(0, 10);
+          if (typeof data.avatarColor === "string" && /^#[0-9a-fA-F]{6}$/.test(data.avatarColor)) {
+            profile.avatarColor = data.avatarColor;
+          }
+          if (typeof data.avatarStyle === "string") profile.avatarStyle = data.avatarStyle.substring(0, 30);
+          if (typeof data.avatarSeed === "string") profile.avatarSeed = data.avatarSeed.substring(0, 30);
+          if (typeof data.avatarCustom === "string") profile.avatarCustom = data.avatarCustom;
+          if (data.avatarCustom === null || data.avatarCustom === "") profile.avatarCustom = undefined;
+          if (users.isMultiUser()) {
+            var mu = getMultiUserFromReq(req);
+            if (!mu) {
+              res.writeHead(401, { "Content-Type": "application/json" });
+              res.end('{"error":"unauthorized"}');
+              return;
+            }
+            users.updateUserProfile(mu.id, profile);
+            // Broadcast updated avatar/presence to all projects
+            projects.forEach(function (pCtx) {
+              pCtx.refreshUserProfile(mu.id);
+            });
+          } else {
+            fs.writeFileSync(profilePath, JSON.stringify(profile, null, 2), "utf8");
+            if (process.platform !== "win32") {
+              try { fs.chmodSync(profilePath, 0o600); } catch (chmodErr) {}
+            }
+          }
+          res.writeHead(200, { "Content-Type": "application/json" });
+          res.end(JSON.stringify(profile));
+        } catch (e) {
+          res.writeHead(400, { "Content-Type": "application/json" });
+          res.end(JSON.stringify({ error: "Invalid request" }));
+        }
+      });
+      return true;
+    }
+
+    // Upload custom avatar image
+    if (req.method === "POST" && fullUrl === "/api/avatar") {
+      var chunks = [];
+      var totalSize = 0;
+      var maxSize = 2 * 1024 * 1024; // 2MB
+      req.on("data", function (chunk) {
+        totalSize += chunk.length;
+        if (totalSize <= maxSize) chunks.push(chunk);
+      });
+      req.on("end", function () {
+        if (totalSize > maxSize) {
+          res.writeHead(413, { "Content-Type": "application/json" });
+          res.end('{"error":"File too large (max 2MB)"}');
+          return;
+        }
+        var raw = Buffer.concat(chunks);
+        // Detect content type from magic bytes
+        var ct = null;
+        if (raw[0] === 0xFF && raw[1] === 0xD8) ct = "image/jpeg";
+        else if (raw[0] === 0x89 && raw[1] === 0x50) ct = "image/png";
+        else if (raw[0] === 0x47 && raw[1] === 0x49) ct = "image/gif";
+        else if (raw[0] === 0x52 && raw[1] === 0x49) ct = "image/webp";
+        if (!ct) {
+          res.writeHead(400, { "Content-Type": "application/json" });
+          res.end('{"error":"Unsupported image format"}');
+          return;
+        }
+        var ext = ct.split("/")[1] === "jpeg" ? "jpg" : ct.split("/")[1];
+        var avatarDir = path.join(CONFIG_DIR, "avatars");
+        fs.mkdirSync(avatarDir, { recursive: true });
+
+        var userId = "default";
+        if (users.isMultiUser()) {
+          var mu = getMultiUserFromReq(req);
+          if (!mu) {
+            res.writeHead(401, { "Content-Type": "application/json" });
+            res.end('{"error":"unauthorized"}');
+            return;
+          }
+          userId = mu.id;
+        }
+        var filename = userId + "." + ext;
+        // Remove old avatar files for this user
+        try {
+          var existing = fs.readdirSync(avatarDir);
+          for (var ei = 0; ei < existing.length; ei++) {
+            if (existing[ei].startsWith(userId + ".")) {
+              fs.unlinkSync(path.join(avatarDir, existing[ei]));
+            }
+          }
+        } catch (e) {}
+        var avatarFilePath = path.join(avatarDir, filename);
+        fs.writeFileSync(avatarFilePath, raw);
+        try { fs.chmodSync(avatarFilePath, 0o644); } catch (e) {}
+        try { fs.chmodSync(avatarDir, 0o755); } catch (e) {}
+        res.writeHead(200, { "Content-Type": "application/json" });
+        res.end(JSON.stringify({ ok: true, avatar: "/api/avatar/" + userId + "?v=" + Date.now() }));
+      });
+      return true;
+    }
+
+    // Serve custom avatar image
+    if (req.method === "GET" && fullUrl.startsWith("/api/avatar/")) {
+      var avatarUserId = fullUrl.split("/api/avatar/")[1].split("?")[0];
+      var avatarDir = path.join(CONFIG_DIR, "avatars");
+      try {
+        var files = fs.readdirSync(avatarDir);
+        var match = null;
+        for (var fi = 0; fi < files.length; fi++) {
+          if (files[fi].startsWith(avatarUserId + ".")) {
+            match = files[fi];
+            break;
+          }
+        }
+        if (match) {
+          var ext = match.split(".").pop();
+          var ctMap = { jpg: "image/jpeg", png: "image/png", gif: "image/gif", webp: "image/webp" };
+          res.writeHead(200, {
+            "Content-Type": ctMap[ext] || "application/octet-stream",
+            "Cache-Control": "public, max-age=31536000, immutable",
+          });
+          res.end(fs.readFileSync(path.join(avatarDir, match)));
+          return true;
+        }
+      } catch (e) {}
+      res.writeHead(404, { "Content-Type": "application/json" });
+      res.end('{"error":"not found"}');
+      return true;
+    }
+
+    // Upload custom avatar for a mate
+    if (req.method === "POST" && fullUrl.startsWith("/api/mate-avatar/")) {
+      var mateIdFromUrl = fullUrl.split("/api/mate-avatar/")[1].split("?")[0];
+      if (!mateIdFromUrl) {
+        res.writeHead(400, { "Content-Type": "application/json" });
+        res.end('{"error":"Missing mate ID"}');
+        return true;
+      }
+      var chunks = [];
+      var totalSize = 0;
+      var maxSize = 2 * 1024 * 1024; // 2MB
+      req.on("data", function (chunk) {
+        totalSize += chunk.length;
+        if (totalSize <= maxSize) chunks.push(chunk);
+      });
+      req.on("end", function () {
+        if (totalSize > maxSize) {
+          res.writeHead(413, { "Content-Type": "application/json" });
+          res.end('{"error":"File too large (max 2MB)"}');
+          return;
+        }
+        var raw = Buffer.concat(chunks);
+        var ct = null;
+        if (raw[0] === 0xFF && raw[1] === 0xD8) ct = "image/jpeg";
+        else if (raw[0] === 0x89 && raw[1] === 0x50) ct = "image/png";
+        else if (raw[0] === 0x47 && raw[1] === 0x49) ct = "image/gif";
+        else if (raw[0] === 0x52 && raw[1] === 0x49) ct = "image/webp";
+        if (!ct) {
+          res.writeHead(400, { "Content-Type": "application/json" });
+          res.end('{"error":"Unsupported image format"}');
+          return;
+        }
+        var userId = null;
+        if (users.isMultiUser()) {
+          var mu = getMultiUserFromReq(req);
+          if (!mu) {
+            res.writeHead(401, { "Content-Type": "application/json" });
+            res.end('{"error":"unauthorized"}');
+            return;
+          }
+          userId = mu.id;
+        }
+        var mateCtx = mates.buildMateCtx(userId);
+        var mate = mates.getMate(mateCtx, mateIdFromUrl);
+        if (!mate) {
+          res.writeHead(404, { "Content-Type": "application/json" });
+          res.end('{"error":"Mate not found"}');
+          return;
+        }
+        var ext = ct.split("/")[1] === "jpeg" ? "jpg" : ct.split("/")[1];
+        var avatarDir = path.join(CONFIG_DIR, "mate-avatars");
+        fs.mkdirSync(avatarDir, { recursive: true });
+        var filename = mateIdFromUrl + "." + ext;
+        // Remove old avatar files for this mate
+        try {
+          var existing = fs.readdirSync(avatarDir);
+          for (var ei = 0; ei < existing.length; ei++) {
+            if (existing[ei].startsWith(mateIdFromUrl + ".")) {
+              fs.unlinkSync(path.join(avatarDir, existing[ei]));
+            }
+          }
+        } catch (e) {}
+        var mateAvatarFilePath = path.join(avatarDir, filename);
+        fs.writeFileSync(mateAvatarFilePath, raw);
+        try { fs.chmodSync(mateAvatarFilePath, 0o644); } catch (e) {}
+        try { fs.chmodSync(avatarDir, 0o755); } catch (e) {}
+        var avatarPath = "/api/mate-avatar/" + mateIdFromUrl + "?v=" + Date.now();
+        // Update mate profile with custom avatar URL
+        var profile = mate.profile || {};
+        profile.avatarCustom = avatarPath;
+        mates.updateMate(mateCtx, mateIdFromUrl, { profile: profile });
+        res.writeHead(200, { "Content-Type": "application/json" });
+        res.end(JSON.stringify({ ok: true, avatar: avatarPath }));
+      });
+      return true;
+    }
+
+    // Serve custom mate avatar image
+    if (req.method === "GET" && fullUrl.startsWith("/api/mate-avatar/")) {
+      var mateAvatarId = fullUrl.split("/api/mate-avatar/")[1].split("?")[0];
+      var mateAvatarDir = path.join(CONFIG_DIR, "mate-avatars");
+      try {
+        var files = fs.readdirSync(mateAvatarDir);
+        var match = null;
+        for (var fi = 0; fi < files.length; fi++) {
+          if (files[fi].startsWith(mateAvatarId + ".")) {
+            match = files[fi];
+            break;
+          }
+        }
+        if (match) {
+          var ext = match.split(".").pop();
+          var ctMap = { jpg: "image/jpeg", png: "image/png", gif: "image/gif", webp: "image/webp" };
+          res.writeHead(200, {
+            "Content-Type": ctMap[ext] || "application/octet-stream",
+            "Cache-Control": "public, max-age=31536000, immutable",
+          });
+          res.end(fs.readFileSync(path.join(mateAvatarDir, match)));
+          return true;
+        }
+      } catch (e) {}
+      res.writeHead(404, { "Content-Type": "application/json" });
+      res.end('{"error":"not found"}');
+      return true;
+    }
+
+    // Change own PIN (multi-user mode)
+    if (req.method === "PUT" && fullUrl === "/api/user/pin") {
+      if (!users.isMultiUser()) {
+        res.writeHead(404, { "Content-Type": "application/json" });
+        res.end('{"error":"Not found"}');
+        return true;
+      }
+      var mu = getMultiUserFromReq(req);
+      if (!mu) {
+        res.writeHead(401, { "Content-Type": "application/json" });
+        res.end('{"error":"unauthorized"}');
+        return true;
+      }
+      var body = "";
+      req.on("data", function (chunk) { body += chunk; });
+      req.on("end", function () {
+        try {
+          var data = JSON.parse(body);
+          if (!data.newPin || typeof data.newPin !== "string" || !/^\d{6}$/.test(data.newPin)) {
+            res.writeHead(400, { "Content-Type": "application/json" });
+            res.end('{"error":"PIN must be exactly 6 digits"}');
+            return;
+          }
+          var result = users.updateUserPin(mu.id, data.newPin);
+          if (result.error) {
+            res.writeHead(400, { "Content-Type": "application/json" });
+            res.end(JSON.stringify({ error: result.error }));
+            return;
+          }
+          res.writeHead(200, { "Content-Type": "application/json" });
+          res.end('{"ok":true}');
+        } catch (e) {
+          res.writeHead(400, { "Content-Type": "application/json" });
+          res.end('{"error":"Invalid request"}');
+        }
+      });
+      return true;
+    }
+
+    // PUT /api/user/auto-continue
+    if (req.method === "PUT" && fullUrl === "/api/user/auto-continue") {
+      var mu = getMultiUserFromReq(req);
+      if (!mu) {
+        // Single-user: use daemon config fallback
+        var body = "";
+        req.on("data", function (chunk) { body += chunk; });
+        req.on("end", function () {
+          try {
+            var data = JSON.parse(body);
+            if (typeof opts.onSetAutoContinue === "function") {
+              opts.onSetAutoContinue(!!data.enabled);
+            }
+            res.writeHead(200, { "Content-Type": "application/json" });
+            res.end(JSON.stringify({ ok: true, autoContinueOnRateLimit: !!data.enabled }));
+          } catch (e) {
+            res.writeHead(400, { "Content-Type": "application/json" });
+            res.end('{"error":"Invalid request"}');
+          }
+        });
+        return true;
+      }
+      var body = "";
+      req.on("data", function (chunk) { body += chunk; });
+      req.on("end", function () {
+        try {
+          var data = JSON.parse(body);
+          var result = users.setAutoContinue(mu.id, !!data.enabled);
+          if (result.error) {
+            res.writeHead(400, { "Content-Type": "application/json" });
+            res.end(JSON.stringify({ error: result.error }));
+            return;
+          }
+          res.writeHead(200, { "Content-Type": "application/json" });
+          res.end(JSON.stringify({ ok: true, autoContinueOnRateLimit: result.autoContinueOnRateLimit }));
+        } catch (e) {
+          res.writeHead(400, { "Content-Type": "application/json" });
+          res.end('{"error":"Invalid request"}');
+        }
+      });
+      return true;
+    }
+
+    // PUT /api/user/chat-layout
+    if (req.method === "PUT" && fullUrl === "/api/user/chat-layout") {
+      var mu = getMultiUserFromReq(req);
+      if (!mu) {
+        // Single-user: save to daemon config
+        var body = "";
+        req.on("data", function (chunk) { body += chunk; });
+        req.on("end", function () {
+          try {
+            var data = JSON.parse(body);
+            var val = (data.layout === "bubble") ? "bubble" : "channel";
+            if (typeof opts.onSetChatLayout === "function") {
+              opts.onSetChatLayout(val);
+            }
+            res.writeHead(200, { "Content-Type": "application/json" });
+            res.end(JSON.stringify({ ok: true, chatLayout: val }));
+          } catch (e) {
+            res.writeHead(400, { "Content-Type": "application/json" });
+            res.end('{"error":"Invalid request"}');
+          }
+        });
+        return true;
+      }
+      var body = "";
+      req.on("data", function (chunk) { body += chunk; });
+      req.on("end", function () {
+        try {
+          var data = JSON.parse(body);
+          var result = users.setChatLayout(mu.id, data.layout);
+          if (result.error) {
+            res.writeHead(400, { "Content-Type": "application/json" });
+            res.end(JSON.stringify({ error: result.error }));
+            return;
+          }
+          res.writeHead(200, { "Content-Type": "application/json" });
+          res.end(JSON.stringify({ ok: true, chatLayout: result.chatLayout }));
+        } catch (e) {
+          res.writeHead(400, { "Content-Type": "application/json" });
+          res.end('{"error":"Invalid request"}');
+        }
+      });
+      return true;
+    }
+
+    // POST /api/user/mate-onboarded
+    if (req.method === "POST" && fullUrl === "/api/user/mate-onboarded") {
+      var mu = getMultiUserFromReq(req);
+      if (!mu) {
+        // Single-user: save to daemon config
+        if (typeof opts.onSetMateOnboarded === "function") {
+          opts.onSetMateOnboarded();
+        }
+        res.writeHead(200, { "Content-Type": "application/json" });
+        res.end('{"ok":true}');
+      } else {
+        users.setMateOnboarded(mu.id);
+        res.writeHead(200, { "Content-Type": "application/json" });
+        res.end('{"ok":true}');
+      }
+      return true;
+    }
+
+    // GET /api/user/auto-continue
+    if (req.method === "GET" && fullUrl === "/api/user/auto-continue") {
+      var mu = getMultiUserFromReq(req);
+      if (!mu) {
+        // Single-user: read from daemon config
+        var enabled = false;
+        if (typeof opts.onGetDaemonConfig === "function") {
+          var dc = opts.onGetDaemonConfig();
+          enabled = !!dc.autoContinueOnRateLimit;
+        }
+        res.writeHead(200, { "Content-Type": "application/json" });
+        res.end(JSON.stringify({ autoContinueOnRateLimit: enabled }));
+        return true;
+      }
+      var val = users.getAutoContinue(mu.id);
+      res.writeHead(200, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ autoContinueOnRateLimit: val }));
+      return true;
+    }
+
+    return false;
+  }
+
+  return { handleRequest: handleRequest };
+}
+
+module.exports = { attachSettings: attachSettings };
